libc: Move com_fs_{exec, apply, undo}_{fchdir, mkstemp}() to comfs.c
