+++ lib/modules/libc/src/fs/comfs.c
@@ -9,8 +9,10 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <sys/stat.h>
 #include <picotm/picotm-module.h>
 #include <unistd.h>
+#include "errcode.h"
 #include "types.h"
 #include "mutex.h"
 #include "counter.h"
@@ -262,3 +264,151 @@ com_fs_finish(struct com_fs *data)
     return 0;
 }
 
+/*
+ * fchdir()
+ */
+
+int
+com_fs_exec_fchdir(struct com_fs *data, int fildes)
+{
+    assert(data);
+
+    /* Reference new directory's file descriptor */
+
+    struct fd *fd = fdtab+fildes;
+
+    int err = fd_ref(fd, fildes, 0);
+
+    if (err) {
+        return err;
+    }
+
+    /* Check file descriptor */
+
+    struct stat buf;
+
+    if ((fstat(fildes, &buf) < 0) || (!S_ISDIR(buf.st_mode))) {
+
+        fd_unref(fd, fildes);
+
+        return ERR_SYSTEM;
+    }
+
+    if (data->newcwd < 0) {
+
+        /* Inject event */
+
+        if (com_fs_inject(data, ACTION_FCHDIR, -1) < 0) {
+            return -1;
+        }
+    } else {
+
+        /* Replace old CWD with new CWD */
+
+        struct fd *fd = fdtab+data->newcwd;
+
+        fd_unref(fd, data->newcwd);
+    }
+
+    data->newcwd = fildes;
+
+    return 0;
+}
+
+int
+com_fs_apply_fchdir(struct com_fs *data, int cookie)
+{
+    assert(data);
+
+    return TEMP_FAILURE_RETRY(fchdir(data->newcwd));
+}
+
+int
+com_fs_undo_fchdir(struct com_fs *data, int cookie)
+{
+    assert(data);
+
+    return 0;
+}
+
+/*
+ * mkstemp()
+ */
+
+int
+com_fs_exec_mkstemp(struct com_fs *data, char *pathname)
+{
+    assert(data);
+    assert(pathname);
+
+    /* Construct absolute pathname */
+
+    char *abspath = com_fs_absolute_path(data, pathname);
+
+    if (!abspath) {
+        return -1;
+    }
+
+    /* Make call */
+
+    int fildes = mkstemp(abspath);
+
+    if (fildes < 0) {
+        free(abspath);
+        return -1;
+    }
+
+    /* Copy trailing filled XXXXXX back to pathname */
+    memcpy(pathname+strlen(pathname)-6, abspath+strlen(abspath)-6, 6);
+
+    free(abspath);
+
+    if (com_fs_inject(data, ACTION_MKSTEMP, fildes) < 0) {
+        return -1;
+    }
+
+    return fildes;
+}
+
+int
+com_fs_apply_mkstemp(struct com_fs *data, int cookie)
+{
+    return 0;
+}
+
+/* Remove temporary file in a quite reliable, but Linux-only, way. This is
+ * only possible because it is certain that the transaction created that file
+ * initially. Note that there is still a race condition. An attacker could
+ * replace the file at 'canonpath' while the process is between stat and
+ * unlink.
+ */
+int
+com_fs_undo_mkstemp(struct com_fs *data, int cookie)
+{
+    char path[64];
+
+    sprintf(path, "/proc/self/fd/%d", cookie);
+
+    char *canonpath = canonicalize_file_name(path);
+
+    if (canonpath) {
+
+        struct stat buf[2];
+
+        if (fstat(cookie, buf+0) != -1
+            && stat(canonpath, buf+1) != -1
+            && buf[0].st_dev == buf[1].st_dev
+            && buf[0].st_ino == buf[1].st_ino) {
+
+            if (unlink(canonpath) < 0) {
+                perror("unlink");
+            }
+        }
+
+        free(canonpath);
+    } else {
+        perror("canonicalize_file_name");
+    }
+
+    return TEMP_FAILURE_RETRY(close(cookie));
+}
